# doc/api.yaml
openapi: 3.0.3
info:
  title: API WASAtext
  description: |-
    API necessarie a WASAText.
  version: "1.0.0"
...
paths:
...
  /wasachat/{utente}/chats/{chat}:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'


      - name: chat
        in: path
        required: true
        description: ID della chat
        schema:
          type: integer
          example: 1
          minimum: 1
          maximum: 1024
    get:
      tags: ["Chat"]
      summary: Ottieni messaggi chat (getConversation)
      description: |-
        Restituisce i messaggi della chat specificata.
      operationId: getConversation
      responses:
        "200":
          description: Messaggi restituiti con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaMessaggiChat'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Chat non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
    post:
      tags: ["Chat"]
      summary: Invia messaggio (sendMessage)
      description: |-
        Invia un messaggio nella chat specificata.
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaMessaggio'
        required: true
      responses:
        "200":
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Chat non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
    delete:
      tags: ["Gestione Gruppi"]
      summary: Esci dal gruppo (leaveGroup)
      description: |-
        L'utente esce dal gruppo specificato.
      operationId: leaveGroup
      responses:
        "200":
          description: Uscita dal gruppo avvenuta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Gruppo non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
  /wasachat/{utente}/risposta/chats/{chat}/messaggi/{messaggio}:
    parameters:
      - name: utente
        in: path
        required: true
        description: Nickname dell'utente
        schema:
          type: string
          example: "Carlo"
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9._~-]+$'

      - name: chat
        in: path
        required: true
        description: ID della chat in cui ci si trova
        schema:
          type: integer
          example: 2
          minimum: 1
          maximum: 1024
      - name: messaggio
        in: path
        required: true
        description: ID del messaggio da inoltrare
        schema:
          type: integer
          example: 2
          minimum: 1
          maximum: 1024
    post:
      tags: ["Chat"]
      summary: Rispondi a messaggio (ansMessage)
      description: |-
        Rispondi ad un messaggio a un messaggio.
      operationId: ansMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaMessaggio'
        required: true
      responses:
        "200":
          description: Risposta aggiunta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio o chat non trovati
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
  /wasachat/{utente}/chats/{chat}/messaggi/{messaggio}:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    - name: chat
      in: path
      required: true
      description: ID della chat
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    - name: messaggio
      in: path
      required: true
      description: ID del messaggio
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    delete:
      tags: ["Chat"]
      summary: Elimina messaggio (deleteMessage)
      description: |-
        Elimina un messaggio dalla chat.
      operationId: deleteMessage
      responses:
        "200":
          description: Messaggio eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/messaggi/{messaggio}:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'

    - name: messaggio
      in: path
      required: true
      description: ID del messaggio
      schema:
        type: integer
        example: 2
        minimum: 1
        maximum: 1024
    post:
      tags: ["Chat"]
      summary: Commenta messaggio (commentMessage)
      description: |-
        Aggiunge un commento a un messaggio.
      operationId: commentMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaCommento'
        required: true
      responses:
        "200":
          description: Commento aggiunto con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
    delete:
      tags: ["Chat"]
      summary: Elimina commento (deleteComment)
      description: |-
        Elimina un commento da un messaggio.
      operationId: deleteComment
      responses:
        "200":
          description: Commento eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Commento non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
components:
  schemas:
...
    RispostaMessaggiChat:
      type: object
      description: Elenco dei messaggi in una chat
      required:
        - messages
      properties:
        messages:
          type: array
          description: Lista dei messaggi
          minItems: 0
          maxItems: 1024
          items:
            $ref: '#/components/schemas/Messaggio'
...
    Messaggio:
      type: object
      description: Dettagli di un singolo messaggio
      required:
        - message_id
        - autore
        - time
        - ricevuto
        - letto
        - commenti
        - inoltrato
      properties:
        message_id:
          type: integer
          description: ID univoco del messaggio
          example: 12
          minimum: 1
          maximum: 1024
        autore:
          type: string
          description: Nickname dell'autore
          example: "Marcolino"
          minLength: 1
          maxLength: 50
          pattern: '^.*$'
        text:
          type: string
          description: Contenuto testuale del messaggio
          example: "Sei libero sabato?"
          minLength: 1
          maxLength: 1000
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
        time:
          type: string
          format: date-time
          description: Timestamp del messaggio in formato ISO 8601
          example: "2025-03-15T14:30:00Z"
          minLength: 20
          maxLength: 30
          pattern: '^.*$'
        ricevuto:
          type: boolean
          description: Indica se il messaggio è stato ricevuto
          example: true
        letto:
          type: boolean
          description: Indica se il messaggio è stato letto
          example: false
        commenti:
          type: array
          description: Lista dei commenti associati al messaggio
          minItems: 0
          maxItems: 100
          items:
            $ref: '#/components/schemas/Commento'
        inoltrato:
          type: boolean
          description: Indica se il messaggio è stato inoltrato
          example: true
        risposta:
          type: integer
          description: ID del messaggio a cui si sta rispondendo
          example: 9
          minimum: 1
          maximum: 1024
    RichiestaMessaggio:
      type: object
      description: Richiesta per inviare un messaggio
      properties:
        testo:
          type: string
          description: Contenuto testuale del messaggio
          example: "Ti va di uscire insieme questo venerdì?"
          minLength: 1
          maxLength: 1000
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    RichiestaCommento:
      type: object
      description: Richiesta per aggiungere un commento
      required:
        - contenuto
      properties:
        contenuto:
          type: string
          description: Contenuto del commento (emoji)
          example: "👍"
          minLength: 1
          maxLength: 1000
          pattern: '^[\u0020-\u1FAFF]+$'
    Commento:
      type: object
      description: Dettagli di un singolo commento
      required:
        - comment_id
        - autore
        - reazione
      properties:
        comment_id:
          type: integer
          description: ID univoco del commento
          example: 2
          minimum: 1
          maximum: 1024
        autore:
          type: string
          description: Nickname dell'autore
          example: "Marcolino"
          minLength: 1
          maxLength: 50
          pattern: '^.*$'
        reazione:
          type: string
          description: Contenuto del commento (emoji o testo breve)
          example: "👍"
          minLength: 1
          maxLength: 10
          pattern: '^[\u0020-\u1FAFF]+$'
...
    RispostaGenerica:
      type: object
      description: Risposta generica
      required:
        - codice
        - risposta
      properties:
        codice:
          type: integer
          description: codice di risposta
          example: 200
          minLength: 1
          maxLength: 1024
          minimum: 200
          maximum: 299
        risposta:
          type: string
          description: risposta 
          example: "messaggio di successo"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    ErroreGenerico:
      type: object
      description: Errore
      required:
        - codice
        - errore
      properties:
        codiceErrore:
          type: integer
          description: codice di errore
          example: 500
          minLength: 1
          maxLength: 1024
          minimum: 300
          maximum: 599
        errore:
          type: string
          description: errore
          pattern: '^.*$'
          example: "errore"
          minLength: 1
          maxLength: 1024